Move trusted provider lookup into identity provider model
